Made more landing pages, updated each to reflect designer and project lead’s wishes
Next Steps:
L3 was selected as the page we use for this project, so need to make folder for the remaining and finish modifying L3 as lead wishes. Build Main page which will resemble ACE-Feature-Video layout, details provided by lead. Create student pages in accordance to wireframes provided by lead.